requirements.txt
pip==9.0.1 # The /environmentDump endpoint has a dependency on this version, as version 10 causes this endpoint to break
lxml==4.2.1
requests==2.18.4
Flask==0.12.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
healthcheck==1.3.2
gunicorn==19.7.1

# For running unit tests
nose==1.3.7
coverage==4.5.1
Flask-Testing==0.7.1

# Get latest commit and tag from GitHub
PyGithub==1.39
